+++ modules/apps/account/account-admin-web/src/main/java/com/liferay/account/admin/web/internal/frontend/taglib/servlet/taglib/ui/AccountEntryOrganizationsScreenNavigationCategory.java
@@ -12,13 +12,12 @@
  * details.
  */
 
-package com.liferay.account.admin.web.internal.frontend.taglib.servlet.taglib;
+package com.liferay.account.admin.web.internal.frontend.taglib.servlet.taglib.ui;
 
 import com.liferay.account.admin.web.internal.constants.AccountScreenNavigationEntryConstants;
 import com.liferay.account.admin.web.internal.display.AccountEntryDisplay;
 import com.liferay.account.admin.web.internal.security.permission.resource.AccountEntryPermission;
 import com.liferay.account.constants.AccountActionKeys;
-import com.liferay.account.constants.AccountConstants;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
 import com.liferay.portal.kernel.exception.PortalException;
@@ -29,42 +28,42 @@ import com.liferay.portal.kernel.model.User;
 import com.liferay.portal.kernel.security.permission.PermissionCheckerFactoryUtil;
 
 import java.util.Locale;
-import java.util.Objects;
 
 import org.osgi.service.component.annotations.Component;
 
 /**
- * @author Albert Lee
+ * @author Pei-Jung Lan
  */
 @Component(
 	property = {
-		"screen.navigation.category.order:Integer=20",
+		"screen.navigation.category.order:Integer=30",
 		"screen.navigation.entry.order:Integer=10"
 	},
 	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
 )
-public class AccountEntryUsersScreenNavigationCategory
+public class AccountEntryOrganizationsScreenNavigationCategory
 	extends BaseAccountEntryScreenNavigationEntry
 	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
-		return AccountScreenNavigationEntryConstants.CATEGORY_KEY_USERS;
+		return AccountScreenNavigationEntryConstants.CATEGORY_KEY_ORGANIZATIONS;
 	}
 
 	@Override
 	public String getEntryKey() {
-		return AccountScreenNavigationEntryConstants.ENTRY_KEY_USERS;
+		return AccountScreenNavigationEntryConstants.ENTRY_KEY_ORGANIZATIONS;
 	}
 
 	@Override
 	public String getJspPath() {
-		return "/account_entries_admin/account_entry/view_account_users.jsp";
+		return "/account_entries_admin/account_entry" +
+			"/view_account_organizations.jsp";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "users");
+		return LanguageUtil.get(locale, "organizations");
 	}
 
 	@Override
@@ -75,18 +74,11 @@ public class AccountEntryUsersScreenNavigationCategory
 			return false;
 		}
 
-		if (Objects.equals(
-				accountEntryDisplay.getType(),
-				AccountConstants.ACCOUNT_ENTRY_TYPE_PERSON)) {
-
-			return false;
-		}
-
 		try {
 			if (AccountEntryPermission.contains(
 					PermissionCheckerFactoryUtil.create(user),
 					accountEntryDisplay.getAccountEntryId(),
-					AccountActionKeys.VIEW_USERS)) {
+					AccountActionKeys.VIEW_ORGANIZATIONS)) {
 
 				return true;
 			}
@@ -101,6 +93,6 @@ public class AccountEntryUsersScreenNavigationCategory
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		AccountEntryUsersScreenNavigationCategory.class);
+		AccountEntryOrganizationsScreenNavigationCategory.class);
 
 }
