+++ modules/apps/account/account-admin-web/src/main/java/com/liferay/account/admin/web/internal/frontend/taglib/servlet/taglib/ui/AccountEntryRolesScreenNavigationCategory.java
@@ -12,20 +12,14 @@
  * details.
  */
 
-package com.liferay.account.admin.web.internal.frontend.taglib.servlet.taglib;
+package com.liferay.account.admin.web.internal.frontend.taglib.servlet.taglib.ui;
 
 import com.liferay.account.admin.web.internal.constants.AccountScreenNavigationEntryConstants;
 import com.liferay.account.admin.web.internal.display.AccountEntryDisplay;
-import com.liferay.account.admin.web.internal.security.permission.resource.AccountEntryPermission;
-import com.liferay.account.constants.AccountActionKeys;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.PermissionCheckerFactoryUtil;
 
 import java.util.Locale;
 
@@ -36,63 +30,44 @@ import org.osgi.service.component.annotations.Component;
  */
 @Component(
 	property = {
-		"screen.navigation.category.order:Integer=30",
+		"screen.navigation.category.order:Integer=60",
 		"screen.navigation.entry.order:Integer=10"
 	},
 	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
 )
-public class AccountEntryOrganizationsScreenNavigationCategory
+public class AccountEntryRolesScreenNavigationCategory
 	extends BaseAccountEntryScreenNavigationEntry
 	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
-		return AccountScreenNavigationEntryConstants.CATEGORY_KEY_ORGANIZATIONS;
+		return AccountScreenNavigationEntryConstants.CATEGORY_KEY_ROLES;
 	}
 
 	@Override
 	public String getEntryKey() {
-		return AccountScreenNavigationEntryConstants.ENTRY_KEY_ORGANIZATIONS;
+		return AccountScreenNavigationEntryConstants.ENTRY_KEY_ROLES;
 	}
 
 	@Override
 	public String getJspPath() {
-		return "/account_entries_admin/account_entry" +
-			"/view_account_organizations.jsp";
+		return "/account_entries_admin/account_entry/view_account_roles.jsp";
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "organizations");
+		return LanguageUtil.get(locale, "roles");
 	}
 
 	@Override
 	public boolean isVisible(
 		User user, AccountEntryDisplay accountEntryDisplay) {
 
-		if (accountEntryDisplay.getAccountEntryId() == 0) {
-			return false;
-		}
-
-		try {
-			if (AccountEntryPermission.contains(
-					PermissionCheckerFactoryUtil.create(user),
-					accountEntryDisplay.getAccountEntryId(),
-					AccountActionKeys.VIEW_ORGANIZATIONS)) {
-
-				return true;
-			}
-		}
-		catch (PortalException portalException) {
-			if (_log.isDebugEnabled()) {
-				_log.debug(portalException, portalException);
-			}
+		if (accountEntryDisplay.getAccountEntryId() > 0) {
+			return true;
 		}
 
 		return false;
 	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		AccountEntryOrganizationsScreenNavigationCategory.class);
 
 }
