+++ modules/apps/account/account-admin-web/src/main/java/com/liferay/account/admin/web/internal/frontend/taglib/servlet/taglib/ui/AccountUserGeneralScreenNavigationCategory.java
@@ -12,39 +12,40 @@
  * details.
  */
 
-package com.liferay.users.admin.web.internal.frontend.taglib.servlet.taglib;
+package com.liferay.account.admin.web.internal.frontend.taglib.servlet.taglib.ui;
 
+import com.liferay.account.admin.web.internal.constants.AccountScreenNavigationEntryConstants;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
 import com.liferay.portal.kernel.language.LanguageUtil;
-import com.liferay.users.admin.constants.UserScreenNavigationEntryConstants;
 
 import java.util.Locale;
 
 import org.osgi.service.component.annotations.Component;
 
 /**
- * @author Pei-Jung Lan
+ * @author Albert Lee
  */
 @Component(
-	property = "screen.navigation.category.order:Integer=20",
+	property = "screen.navigation.category.order:Integer=10",
 	service = ScreenNavigationCategory.class
 )
-public class UserContactScreenNavigationCategory
+public class AccountUserGeneralScreenNavigationCategory
 	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
-		return UserScreenNavigationEntryConstants.CATEGORY_KEY_CONTACT;
+		return AccountScreenNavigationEntryConstants.CATEGORY_KEY_GENERAL;
 	}
 
 	@Override
 	public String getLabel(Locale locale) {
-		return LanguageUtil.get(locale, "contact");
+		return LanguageUtil.get(locale, getCategoryKey());
 	}
 
 	@Override
 	public String getScreenNavigationKey() {
-		return UserScreenNavigationEntryConstants.SCREEN_NAVIGATION_KEY_USERS;
+		return AccountScreenNavigationEntryConstants.
+			SCREEN_NAVIGATION_KEY_ACCOUNT_USER;
 	}
 
 }
